tidy up msgs definitions
